FPM-31.Create user repo and enum Role
