+++ src/main/java/com/ecomerceApi/Priscila/model/Role.java
@@ -1,0 +1,6 @@
+package com.ecomerceApi.Priscila.model;
+
+public enum Role {
+    USER,
+    ADMIN
+}
