--- src/main/java/com/ecomerceApi/Priscila/models/User.java
@@ -1,56 +1,0 @@
-package com.ecomerceApi.Priscila.models;
-
-import jakarta.persistence.*;
-
-@Entity
-@Table
-public class User {
-   private int id;
-    private String name;
-    private String email;
-    private String password;
-    private String role;
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    @Column(unique = true)
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getRole() {
-        return role;
-    }
-
-    public void setRole(String role) {
-        this.role = role;
-    }
-}
